Support for visual range with `:Isort`
This would make the extra Python functions unnecessary, but I've left
them, as with the old mapping for visual mode.

Feel free to improve upon. I would also like to not have the mapping
(i.e. a setting for it). Given how easy `:Is<tab><cr>` is (and the
number of times you use it), it should not overwrite an existing
default Vim shortcut.

/cc https://github.com/fisadev/vim-isort/pull/3